requirement.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.3.0
sqlparse==0.4.4
